move "SetupWorkerApi" and "SetupServerApi" into own glossaries
