+++ src/setupWorker/setupWorker.ts
@@ -1,29 +1,13 @@
-import { SetupWorkerInternalContext, RequestHandlersList } from './glossary'
+import {
+  SetupWorkerInternalContext,
+  RequestHandlersList,
+  SetupWorkerApi,
+} from './glossary'
 import { createStart } from './start/createStart'
 import { createStop } from './stop/createStop'
 import * as requestHandlerUtils from '../utils/handlers/requestHandlerUtils'
 import { isNodeProcess } from '../utils/internal/isNodeProcess'
 import { ServiceWorkerMessage } from '../utils/createBroadcastChannel'
-
-export interface SetupWorkerApi {
-  start: ReturnType<typeof createStart>
-  stop: ReturnType<typeof createStop>
-
-  /**
-   * Prepends given request handlers to the list of existing handlers.
-   */
-  use: (...handlers: RequestHandlersList) => void
-
-  /**
-   * Marks all request handlers that respond using `res.once()` as unused.
-   */
-  restoreHandlers: () => void
-
-  /**
-   * Resets request handlers to the initial list given to the `setupWorker` call, or to the explicit next request handlers list, if given.
-   */
-  resetHandlers: (...nextHandlers: RequestHandlersList) => void
-}
 
 /**
  * Configures a Service Worker with the given request handler functions.
